Once again shifted strategy, but hey it works now.
Switched from QLabel to QPainter to display the frame image because it
seemed to simplify scaling and centering the image and I didn't have to
deal with layouts.

Added an overridden mouseMoveEvent for drag-drawing functionality
because the old way froze the event loop and broke the program, oops.
Also made sure that drag-drawing doesn't emit pixelClicked if the user
drags off of the canvas.

Connected the pixelClicked signal from CanvasPane to the onPixelClicked
slot from Project.

Tested a white default frame and it correctly displayed in the canvas
pane and the pixelClicked signal correctly emitted when clicking or
click-dragging on the canvas.