Real time drawing and a transparency background.
CanvasPane now keeps a reference to the current frame and redraws the
canvas using that reference on clicks and drags (the only time the frame
would change) so now the frame is updated in real time as the user
draws.

Also fixed a bug where invalid coordinates and (0,0) were being handled
the same way so (0,0) wasn't drawing.

Added a png-style background grid to represent the transparent parts of
the sprite. The grid scales with the sprite so that each pixel on the
grid is an empty pixel in the sprite.

Set default pencil color to blue in the model constructor, will remove
before merge.