PreviewPane implemented and it miraculously works.
PreviewPane is constructed with a pointer to the project, from which it
grabs a reference to the sprite and holds it in a member pointer, that
should maybe be made const in the future.

Use a QTimer to time out drawing signals, and used a QPainter to paint
each frame of the sprite, same logic as CanvasPane (like exactly the
same).