Save frame rate int the Sprite object.
Added a signal in previewpane.cpp that emits on frame rate change and is
received by the project model which sets the frame rate held by the
sprite. Frame rate is the ms duration that the frame is displayed for,
in the form 1000/fps = frameRate. Note that frameRate is an int because
QTimer's interval is an int so we just match the timer interval.